Changed unsigned int to uint16_t

+++ Autopilot/AttitudeManager/AttitudeManager.c
@@ -674,8 +674,8 @@ void readDatalink(void){
                 setGains(GSPEED, CMD_TO_FLOAT_ARRAY(cmd->data));
                 break;
 	    case CALIBRATE_PWM_INPUTS:
-	        unsigned int chn_numb = CMD_TO_INT(cmd->data);
-	        unsigned int offset = CMD_TO_INT(cmd->data + 2);
+	        uint16_t chn_numb = CMD_TO_INT(cmd->data);
+	        uint16_t offset = CMD_TO_INT(cmd->data + 2);
 	        float scale = CMD_TO_FLOAT(cmd->data + 4);
 	        calibratePWMInputs(channel, scale, offset);
 	        break;
